contract note delete

+++ app/GraphQL/Mutations/DeleteContractNote.php
@@ -2,7 +2,9 @@
 
 namespace App\GraphQL\Mutations;
 
+use App\Models\Note;
 use Illuminate\Auth\AuthManager;
+use Nuwave\Lighthouse\Exceptions\AuthorizationException;
 
 class DeleteContractNote {
     public function __construct(
@@ -18,13 +20,29 @@ class DeleteContractNote {
      * @return  array<string, mixed>
      */
     public function __invoke($_, array $args): array {
-        $user = $this->auth->user();
-        $note = $user->notes()->whereKey($args['input']['id'])->first();
+        return [
+            'deleted' => $this->deleteNote(
+                $args['input']['id'],
+                ['org-administer', 'contracts-view', 'customers-view'],
+            ),
+        ];
+    }
+
+    /**
+     *
+     * @param array<string> $permissions
+     */
+    public function deleteNote(string $noteId, array $permissions): bool {
+        $note = Note::whereKey($noteId)->first();
+
+        if (!$this->auth->user()->canAny($permissions, [$note])) {
+            throw new AuthorizationException();
+        }
 
         if ($note) {
             $note->delete();
         }
 
-        return ['deleted' => (bool) $note];
+        return (bool) $note;
     }
 }
